re-structure the project so it actually works issue-47 issue-70
