+++ src/hydrology-render.py
@@ -11,7 +11,7 @@ from rasterio.transform import Affine
 import time
 import math
 
-from lib import DataModel, SaveFile, Math
+from lib import ShoreModel, HydrologyNetwork, SaveFile, Terrain, TerrainHoneycomb, Math, TerrainHydrology
 
 parser = argparse.ArgumentParser(
     description='Implementation of Genevaux et al., "Terrain Generation Using Procedural Models Based on Hydrology", ACM Transactions on Graphics, 2013'
@@ -82,13 +82,16 @@ longitude = float(args.longitude)
 db = SaveFile.openDB(inputFile)
 resolution = SaveFile.getResolution(db)
 edgeLength = SaveFile.getEdgeLength(db)
-shore: DataModel.ShoreModel = DataModel.ShoreModel()
+shore: ShoreModel.ShoreModel = ShoreModel.ShoreModel()
 shore.loadFromDB(db)
-hydrology: DataModel.HydrologyNetwork = DataModel.HydrologyNetwork(db)
-cells: DataModel.TerrainHoneycomb = DataModel.TerrainHoneycomb()
-cells.loadFromDB(resolution, edgeLength, shore, hydrology, db)
-Ts: DataModel.Terrain = DataModel.Terrain()
+hydrology: HydrologyNetwork.HydrologyNetwork = HydrologyNetwork.HydrologyNetwork(db)
+cells: TerrainHoneycomb.TerrainHoneycomb = TerrainHoneycomb.TerrainHoneycomb()
+cells.loadFromDB(edgeLength, db)
+Ts: Terrain.Terrain = Terrain.Terrain()
 Ts.loadFromDB(db)
+terrainSystem = TerrainHydrology.TerrainHydrology(edgeLength)
+terrainSystem.hydrology = hydrology
+terrainSystem.cells = cells
 
 radius = edgeLength / 3
 rwidth = edgeLength / 2
@@ -152,7 +155,7 @@ def TerrainFunction(point: typing.Tuple[float,float]) -> float:
     wi=wt_ # IDK why he converts these here
     hi=ht_
     
-    nodeID = cells.nodeID(point)
+    nodeID = terrainSystem.nodeOfPoint(point)
     if nodeID is None:
         return hi
     node = hydrology.node(nodeID)
@@ -197,7 +200,7 @@ octaves = 6
 persistence = 0.5
 lacunarity = 2.0
 # Height of a terrain primitive
-def ht(p: typing.Tuple[float, float], t: DataModel.T) -> float:
+def ht(p: typing.Tuple[float, float], t: Terrain.T) -> float:
     return t.elevation# +pnoise2(p[0]/scale,p[1]/scale,octaves=octaves,persistence=persistence,lacunarity=lacunarity,repeatx=shore.shape[0],repeaty=shore.shape[1],base=0)*10
 
 # Height of a river primitive?
